+++ src/watch/index.js
@@ -256,7 +256,3 @@ export async function getWatchStatus({ repoUrl, token }) {
     return null;
   }
 }
-// license test Tue Jun 23 15:30:36 CDT 2026
-// license JWT test Tue Jun 23 15:48:23 CDT 2026
-// 9.0.6 team tier test Tue Jun 23 16:11:39 CDT 2026
-// 9.0.7 tier test Tue Jun 23 16:36:04 CDT 2026